requirements.txt
python-telegram-bot
pybluez
gattlib